Pass the original user_data to the deleter function
The callback which is used for feeding libyang with module source code
accepts a generic void* for contextualization. The callback itself can,
when called, specify another function which can be used for deleting the
result. The function signature matches that of `free`. However, when
this custom deleter is called, the user_data context is not passed.

This means that the first callback cannot really provide any
contextualization. It could still swap a generic `free` for something
else, but that "something else" again lacks any context because it can
only be a free-standing function.

In the C++ API for this, I will need a real context even for the custom
deleter, otherwise it is not possible to use basic features such as a
std::function.

The Python bindings appear to ignore the custom deleter. I have no clue
if that behavior is correct or if it leaks.

This will require follow-up patches to libnetconf2/Netopeer2-server.