C++-ish way of reading module data via callbacks
This was not possible to do when using C++ bindings because the ly_ctx
was private. Also, while C-style freestanding functions certainly work,
this API can be used more easily I think.

The Python bindings already contained support for something like this. I
do not know if it's possible to reimplement Python's usual semantics on
top of the API introduced in this patch.

Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>