Change constness of the data returned by module implementation callbacks
libyang itself won't modify these data, so a `const char *` is a better
match. That will allow returning, e.g., string literals or C++'s
`std::string::c_str` without any extra deleter.

The cost of this patch is an extra cast which is needed when calling a
deleter. If a deleter is used, then quite obviously that data cannot be
const.