Fix misaligned when pointer in uses/choice/case
Fixes https://github.com/sysrepo/sysrepo/issues/1272.

The root cause of https://github.com/sysrepo/sysrepo/issues/1272 is that
`resolve_applies_when()` erroneously returns true for the ietf-snmp leaf
"use-prefix". `resolve_applies_when()` incorrectly evaluates the parent
container "tsm" to have a when-condition.

The node type for the "tsm" container is `LYS_CONTAINER`.
`resolve_applies_when()` casts the schema node to `struct lys_node_uses *`,
but this struct has a different offset for the `when` pointer than
`lys_node_container`. When the lyb file format was implemented, a 4-byte hash
was added to `lys_node_container` prior to its when pointer, but a
corresponding 4-byte value was not added to `lys_node_uses`. Thus the `when`
pointer in the two structures is now at a different offset. When
`resolve_applies_when()` reads what it thinks is the `when` pointer on the
container node, it is actually reading part of the hash array, which
happens to be non-zero, and so the method thinks there is a when condition
present.

I fixed this by adding the 4-byte hash array to `lys_node_uses`,
`lys_node_choice`, and `lys_node_case`, just above the `when` pointer.