Fix big endian issues writing and reading .LYB files
The print and parse routines which handle .LYB files assume a little endian
CPU. This means libyang is failing when it is compiled for a big endian
target.

For example consider the code:
 lyb_write(out, (uint8_t *)&mod_count, 2, lybs);

mod_count is an integer and on a typical 32 bit big endian target its
format in memory is (byte0(MSB) byte1 byte2 byte3(LSB)) meaning
if mod_count has a value of 1 the first three bytes of mod_count
are zero, and as the routine only write the first two bytes, it writes
out 0. This effectively corrupts the .LYB file.

This patch modifies the routines to write and read .LYB files to use
endian agnostic methods to access data in a .LYB file.