parser BUGFIX prevent use after free
When parsing a "bits" data node, the existing value of the node is first
"freed". Then, the new value is parsed and the data node is updated with
it.

When the parsing fails (because the value is invalid), the leaf node is
left with a "freed" value. This leads to random crashes.

Only free the old value if the parsing succeeds.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>