schema compile BUGFIX unprefixed nodeids in leafref path
Unprefixed nodeids in leafref path are supposed to be mapped to the
module where the leafref type is instantiated, not where the path is
defined (in contrast to the prefixed nodeids). Also the current()
function in predicates does not refer to the node of the predicate, but
the node where the leafref type is instantiated.