schema compile CHANGE do not compile imported schemas on import
When importing schema into a module being compiled, only the parsed
version of the imported module is needed. The compiled version (and then
making the imported module implemented, if needed) is necessary only
when the nodes from the imported module are referenced via augment,
deviate or leafref path. Handling such a situation will be added later,
currently it is not implemented yet.