Remove unused members
Looking at the source, both of these flags appear to be used just
internally within libyang. They are only set and never read, and
therefore clang warns about their use:

 swig/cpp/src/Tree_Data.hpp:89:13: warning: private field 'flags' is not used [-Wunused-private-field]
 swig/cpp/src/Tree_Schema.hpp:434:13: warning: private field 'flags' is not used [-Wunused-private-field]

If they are useful for users of libyang and if they should be made
available, then an accessor and presumably also some consumers of this
variable should be added.