VERSION bump to version 1.0.0
Moving to 1.0 is connected with change of versioning. The version is
now split to library version connected strictly with the libyang library
and project version providing general version of the project. It is
expected that starting with version 2.0 (currently in branch libyang2),
the library version will move forward faster than the project version.