YANG parser BUGFIX improve checking of input data when parsing
Despite the testcase is connected with special UTF8 character, it is
actually detected as unexpected end-of-input since the character starts
with a NULL-byte. In such a case libyang is not able to detect that
after the NULL-byte are still some data to read, it just interpret it as
end of input string.

Other, not-supported (by RFC) UTF8 characters should be detected
correctly with different error message.

Fixes #780