data types CHANGE simplify data types plugins API
All callbacks are now mandatory in built-in types (so no check for the
callback presence is needed now). Also the validate and store callback
are now connected in a store callback which is (according to the
options) supposed to validate, canonize and store data. This change is
needed in the case the validation process needs to compare the current
data with some reference node,