data types BUGFIX validationg instance-identifier with a complete data tree
Correct resolving multiple predicates on lists with multiple keys. When
a later predicate fails to match the current list instance and we are
moving to another instance, it is necessary to reset all the preceidings
predicates and validate them again. So far we were just continuing with
the current (and later) predicate.