JSON parser BUGFIX maintaining hash tables with record of keyless list
Hash of a keyless list was added into the parent's hashtable after
processing all its children. But when the children are parsed, it
leads to an update of the keyless list's hash and try to update also the
parent's hash-table. However, because the record of the keyless list is
not yet stored there, assertion is raised.

For the keyless lists, the hash is now inserted into the parent's
hash-table before procesinng any children and updated with each hash
change.

Fixes #867