schema CHANGE remove context member from lysc_must
In contrast to when statement, which can appear in uses, compiled must is
always placed in its context node, so getting the context node is quite
straightforward and a separate context meber is not necessary.

The idea to inherit must from RPC's input/output to the their data nodes
(and then need the context memebr in lysc_must) is not good since there are
problematic corner cases with augmenting and deviating input/output.