context BUGFIX searching for the latest revision
when the latest revision is already in the context and searching via
import callback and searchapths provides an older revisions, the newest
revision from the context must be used. On the other hand, when some
revision is present in the context (and callback/searchpaths were not
used to get it), the callback/searchpaths should be tried to get a newer
revision.