Introduce API tracing for debug logging
The logs implemented so far simply show function entries to
support debugging and discovery of API usage patterns.

The relevant bit flag is LY_LDGAPI (0x20) and usage is similar to
the existing debugging flags:

        ly_verb(LY_LLDBG);
        ly_verb_dbg(0x20);

Related logs show a "API: " prefix.

So far every function with the 'API' macro is traced with the
exception of logging and printing related API functions.