schema parsers BUGFIX creating implicit cases when augmenting choice
In the case of an implicit case (data node in the choice is specified
without the case statement), libyang is supposed to add the case node to
avoid problems with addressing that implicit case or anything inside it
by another augments or anything else.

This fixes the bug that the implicit case was not added for the data
nodes added to a choice by augment.

Fixes #889