schema parsers BUGFIX duplicating iffeature expression data
When instantiating a grouping into a uses statement, if-feature
expression of the instatiated data nodes are duplicated. There was a bug
when counting the size of the original expression data so in case of a
long expression only a part of the expression was duplicated which led
to a segfault when the duplicated expression was parsed.

Fixes #887