printer: ignore EAGAIN and EINTR errno in callback
When using libnetconf2 via the UNIX or FD transport and sending large
configs, we can get the following error from libyang:

  Print error (Resource temporarily unavailable).

Do not fail when errno holds EAGAIN or EINTR. These are not errors. This
can occur when using ly_print_clb with a callback that does I/O on
a non-blocking file or if the application received a signal during the
write syscall.

If the callback function returns a zero or positive value, reset errno
so that LY_PRINT_RET does not return an error.

Fixes: 609d708ca567 ("printer CHANGE use errno to check for printer errors")
Signed-off-by: Robin Jarry <robin.jarry@6wind.com>