context CHANGE avoid error messages when searching for latest revision import
When searching for the latest revision import (import without explicit
revision) and the context already contains some revision of the required
module, libyang searches for the latest revision in the searchpatchs and
via callback. In the case no revision of the module found in the
searchpaths, an error message was printed despite libyang actually has a
revision of the module in the context and it is going to use it (and
actually mark it as the latest available revision).

So this patch avoids such an error message in this particular case.