schema parsers BUGFIX irefining iffeature expression
When refining a grouping's node in a uses statement, if-feature expression is
duplicated from refine into the target node. There was a bug when counting the
size of the original expression data so in case of a long expression only a
part of the expression was duplicated which led to a segfault when the
duplicated expression was parsed (e.g. when printing tree).

Fixes #962