schema compile BUGFIX handling errors when compiling data node
Different parts of the lys_compile_node() had different expectation
about the previous value of the ret variable, which is returning at the
end and indicates error. The initial value was an error code, but during the
processing it usually got a success code while the following code was
still expecting and error code as a default value for the variable.

At the end it led to accessing freed memory by executing a code which
was not supposed to be executed because of the previous (hidden) error
when compiling the schema.

Fixes #1043