common BUGFIX strict aliasing of sized arrays
Sized array is commonly used for storing array of pointers or other
structures. Each sized array is preceided by a counter providing the
size of the array. Since we used uint32_t type for the counter, pointers
were misalligned on 64b architectures.

Currently, I don't have any performance measurement, but the sized
arrays are supposed to be used quite intensively in libyang2 and the
strict aliasing could have impact on the performance. The cost is the
increased size of each array by 4 additional bytes.

The patch changes all the macros working with the counter to use a
generic LY_ARRAY_SIZE_TYPE. For now, it is set to be uint64_t. In future
it is possible to simply change the type according to the detected
architecture in build time if such a mechanism will make sense.

Fixes #1044