schema BUGFIX parsing identifiers matching keywords
In some case, the schema can have identifier on the same place as a
keyword (prefixed extension instance). In the case the beginning part of
the identifier is matching a keyword (a prefix identityone is matching
identity keyword), it was wrongly detected as an invalid keyword instead
of trying to resolve it as an extension's prefix.