bindings CHANGE add new python CFFI bindings
This is the integration of https://github.com/rjarry/libyang-cffi In the
long term, it is intended to replace the existing SWIG bindings.

Some highlights:

* Uses CFFI for interaction with libyang.so.
* More "pythonic" API. Should be easier to use for Python developers.
* All high-level code is in Python. Should ease maintenance.
* Virtualenv/pip friendly.
* Schema diff feature.
* New "print" data format: python dict.

To enable the generation, add the -DGEN_PYTHON_CFFI_BINDINGS=ON cmake
option. It does not depend on SWIG nor on the CPP bindings. It may be
enabled even if -DGEN_LANGUAGE_BINDINGS is OFF.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>