travis BUGFIX install fixed versions of codecov and twine
This fixes two problems seen during deployment of build artifacts:

The first one is:

 ~# codecov
 ...
 Error running `gcov -pb
 /home/travis/build/CESNET/libyang/build/CMakeFiles/compat.dir/compat/compat.c.gcno`:
 [Errno 2] No such file or directory
 Error: 'NoneType' object has no attribute 'replace'

Which is a bug introduced in a recent version of codecov. Use a fixed
version that we know works.

And the second one:

 ~# twine upload --skip-existing build/python/dist/*
 Uploading distributions to https://upload.pypi.org/legacy/
 TypeError: expected string or bytes-like object

Which seems related to specific a metadata 2.1 field in setup.py:
"description_content_type". The version of twine provided by ubuntu
18.04 is too old for this. Install a recent version of twine with pip to
fix the problem.

Use pip3 for both codecov and twine as pip (python 2) may have problems
with recent versions of python packages.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>