python BUGFIX only include required files in source archive
When building into a subdirectory of the root of the repository, a lot
of generated build files are included into the source archive (intended
for upload on pypi.org). This makes the archive bigger and these files
are not required.

Be more restrictive about files that are included into the source
archive.

Fixes: 6e6475d8d5ad ("bindings CHANGE add new python CFFI bindings")
Signed-off-by: Robin Jarry <robin.jarry@6wind.com>