python BUGFIX remove SNode.parse_data_dict
This method may not work in all cases. For example, if the schema node
object is nested into a list, information about the actual list element
instance in which the data node should be created is missing.

Remove this method and "replace" it with DNode.merge_data_dict() which
will work in all cases.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>