schema compile FEATURE ordered augment connect
When there are augments from several modules
to one node, they are ordered based on their
module name. The point is only to make sure
the augments are always connected in the same
order no matter the order of loading the
modules. That is required for LYB hashes to
work as they should.