schema compile BUGFIX freeing schemas being compiled as a dependency
As a reaction to a compile error, all the schemas previously compiled as
a dependency are supposed to be reverted back to the not-implemented
state. Freeing such schemas-compiled-by-dependency must be done
carefully since some of them could actually cause compilation of the
module which failed and is going to free the dependency modules. We use
current context's set id to identify which modules are safe to free and
which will be freed later when the call stack will return to them.