yanglint BUGFIX handling context search paths
yanglint adds the location of the file to parse as a temporary
searchpath and it is necessary also to check that it is not a
duplication of the already set searchpaths.

In non-interactive mode, there was also a bug when removing this
temporary path since there was used a wrong path index to remove.