Revert "schema CHANGE add summarizing error messages when module parsing/compiling failed"
This temporarily reverts commit 27636e0fef8199c9188691dbc1049025e132e8ea.

There is a need for complex changes in tests. With the summarizing error
message it is necessary to check for multiple error messages which is
not currently supported. Also since this change will affect most of the
schema tests, which needs some refactoring, it is useful to make these
changes together.