fuzz TEST add regression testing for previous issues found by fuzzing
add the inputs that caused the issues to tests/fuzz/corpus directories
for the corresponding fuzz target. This includes a fuzz_regression_test
binary which is used to set up a test for every fuzzing target, wrapping
the fuzz harness and reusing it as a part of the test. The appropriate
corpus is then sent into the fuzz harness to confirm that it doesn't
cause crashes or other similar issues anymore.