fuzz BUGFIX fix build issues with fuzz regression tests
This fixes multiple fuzz_regression_test issues.
First, if libyang was built with both ENABLE_BUILD_TESTS and
ENABLE_FUZZ_TARGETS, the fuzz harnesses wouldn't build as the test fuzz
corpus was copied into a directory of the same name, so a _fuzz postfix
was added to the fuzz harness executables.
The other fixes fix some build warnings and typos.