JSON data parser BUGFIX checking empty input
Checking empty input was done in a different way than it is done in XML
data parser. JSON data parser did the check before even initiating JSON
parser. The code in higher levevel did took it into account so it
might crash.

The patch unifies the way of checking empty input with XML - the lower
(in this case) JSON parser is used in anyway and preserved for higher
level code.