schema tree CHANGE remove lys_find_node() from the API
Hide use of schema path format from the API. The schema path format is
used only in YANG's augments and deviations. We have decided to have
only the data path and XPath formats in the API to keep it as simple as
possible. The cost of this simplicity is that it is not possible to
address structure schema nodes (choices, cases, ...) which cannot be
instantiated in data tree. However, we believe that the real world cases
are about the data nodes.

The schema path is used internally and there are functions working with
this format, so in case there will be need to support it in the API, it
can be added there later.