set CHANGE replace LY_SET_OPT_USEASLIST by a single flag in ly_set_add()
There was just a single option and the ly_set API is quite stable for a
long time, so there will be hardly any new option. So instead of an
option bitmask with a single value, make the ly_set_add() parameter only
a single flag to switch off checking (and ignoring) items already
present in the set.