schema tree CHANGE hide LYSC_OPT_ options from public API
The options are use by lys_compile(), but the function is hidden inside
the schema parsers and there is currently no API (nor useful option) to
pass the options into it. All the options (except LYSC_OPT_FREE_SP) are
used only internally to distinguish some parts of the schemas (RPCs,
groupings, etc.).

The patch also removes LYSC_OPT_FREE_SP since the feature connected with
this option was not finished.

All the LYSC_OPT_ features were also renamed to LYS_COMPILE_ to fit
better with the function where used.