Don't correct DST when parsing dates
If the tm_isdst field doesn't correspond to the date fields, mktime()
tries to correct that by adding one hour to the time and changing the
tm_isdst field. Because the user's local time is being, the DST might
differ from what is being parsed and this type of correction can happen.
If the user's localtime is between 23:00 and 23:59, that means that the
day is also going to change and that breaks the algorithm, because it
depends on the day being the same after normalization.

-1 in tm_isdst means that mktime shouldn't correct the time and just
fill in tm_isdst based on what date says.