log CHANGE unify API of logger control functions
There are 3 functions to control different options of the logger -
ly_log_level(), ly_log_dbg_groups() and ly_log_options(). It is
necessary to unify their usage to make API simple.

Following the umask() behaviour, all the functions now returns the
previous value of the set options. To get the current value, it is
necessary to call the functions twice resetting the set value.

In case of the logger's debug groups, the groups which are not used in
the code of libyang 2.0 were removed.