log FEATURE maintain data for printing error location information
Provide storage for line, schema node, data node, whatever which can
provide useful information where an error ocures. The storage can be
continuously updated about the location where the parser or any other
libyang part works and logger can use it to print all the available
information about the error. It also simplifies the logging
functions/macros since the location information is provided implicitely.

The storage is a thread specific data connected with the context. This
way the storage relates to the error records which are also stored in
the context and the storage can be also easily cleaned up when the
context is destroyed.