schema parsers BUGFIX handling includes in submodules
The result of parsing includes in submodule was incorrectly checked.

Instead of parsing submodule all the times, check the submodule presence
in main module and re-use it.

In YANG 1.0, main module is not required to include all the submodules.
Some of the submodules can be included via another submodule. In such a
case, the include record is now injected also into the main module to be
simply available and to unify it with YANG 1.1. Such a record is marked
as injected and it is not printed.

On the other hand, YANG 1.1 requires all the submodules to be present in
main module. So the situation described above raises error.

Fixes #1353