schema CHANGE move priv field back to common location
There's a setter but not a getter for private data, which makes it ugly
to get the value in client code. But beyond that there's existing code
that assumes this pointer is going to be in the same location (i.e., one
can pull it from a common lys_node, now lysc_node), b/c that's how it
worked in libyang1. Let's not make extra work and debugging for everyone
and leave the priv accessible from the same memory offset as was the
case with libyang1.