Support for parsing JSON fragments and bare leaf values
Add support for parsing JSON fragments by adding a new flag,
LYD_OPT_FRAGMENT, and using it accordingly in lyd_parse_json. The
semantics of this flag are to make use of the passed-in node to start
where the schema nodes are compared against.

Similarly, add support for parsing bare values for top-level (from the
parser's point of view) leaf nodes, using a new LYD_OPT_BARETOPLEAF
flag. The parsing of leaf nodes is too tied in with the allocation of
a new node and parsing of the node name and namespace to easily
trigger it from that case, so instead the case is check and handled
explicitly in the top-level parser loop.