schema BUGFIX processing extension instance argument
Due to differences in YANG and YIN, the extension instance's argument is
processed differently. So far, the argument was missing when the source
format was YIN and extension was not compiled. This was breaking
printers and limiting even other use cases of the argument.

This patch solves the problem by getting the extension definition when
needed and finishing the argument processing not finished by YIN parser.