extensions CHANGE CHANGE cleanup API of the schema printer for extension plugins
Be more clear by what API of the compiled schema printer is available for the
extension plugins. While the printer context is needed, we want to keep its
members hidden to be able to make future changes. Therefore, there is only a
(possibly extensible in a future) set of getters to access the context's members
to be read and updated by the plugins' printer callbacks.

The patch also explicitely states the common compatible members of the
different schema printers.