yanglint BUGFIX freeing output handler on print command
checking for content of output handler had a side effect of printing
error message when the stdout was not inserted and some external
filepath was used. S othere is a need of a flag if the default stdout
was used or there is a filepath-based output handler which need cleanup.