trees REFACTOR avoid implicit casts in macros available from public headers
C++ compilers do not like implicit cast (connected with void* and
allowed in C), and since the public headers are supposed to be usable
also by C++ code, refactor the code to avoid them.

The change is also connected with the de-duplication of the code, so 3
new generic macros are introduced: LY_ARRAY_NEW, LY_ARRAY_CREATE and
LY_LIST_NEW().