extension FEATURE parsed tree in lysp_ext_instance
A pointer to the parsed tree is now available in the lysp_ext_instance
structure. Now it is not necessary to temporarily create a parsed tree
in the lys_compile_extension_instance function. The callback function
lyplg_ext_compile_clb had to lose the const qualifier for the
lysp_ext_instance parameter.