data tree REFACTOR allow lyd_value to be variably sized
To gain performance from plugin types its better for the plugin to be
able to store it's value directly in lyd_value rather than having to
allocate another structure with another pointer-deref and another
cache-line load for each access.

The simple changes included here allow for lyd_value to be variable
sized, which allows for the above.