plugins types REFACTOR most types use lyd_value directly
Added a #define to specify the space reserved in lyd_value for plugin
types specific values. The default (24) is enough to hold most
types (ipv6-address/ipv6-prefix being the largest). This is 2 extra
pointers of space in an already largish structure (if counting
lyd_node_term that holds it).

The #define and macros still allow for miniaturization for people
wanting to do embedded applications where every octet might count.