schema tree REFACTOR dep sets creation
Some idioms should now be always satisfied
such as no module is in 2 dep sets. Also,
dep sets are not broken at a module that
does not have statements for compilation
or recompilation because there still can
be at least groupings that actually may
extend the dep set with its imports.