plugin types BUGFIX ::ly_err_new() format emphasis
This commit prevents an error where err_msg could contain a conversion
specifier (character '%'), but the variadic arguments are empty, which
causes undefined behavior. Therefore, in the function, the err_msg
parameter has been renamed to err_format to emphasize that the function
expects a format. A compiler attribute has also been added to the
function to check the correct number of variadic arguments.