parser json BUGFIX decrease sublevels when skipping empty arrays and objects in lydjson_data_skip
If lydjson_data_skip encountered an open object with a nested empty
object it would increment the sublevel when the empty object was opened,
but wouldn't decrement the sublevel when it was closed, as it only
checked for LYJSON_OBJECT_CLOSE and not LYJSON_EMPTY_OBJECT.

This could result in lydjson_data_skip continuing to parse the input
resulting in invalid state in its caller.

The same issue would happen with empty arrays.

This issue was found by OSS-Fuzz in lyd_parse_mem_json.

Signed-off-by: Juraj Vijtiuk <juraj.vijtiuk@sartura.hr>