build: Require C11
This project uses anonymous structs and anonymous unions which are not
part of C99, but are available since C11. By default, compilers tend to
enable these, but they backfire when building with `-pedantic` -- and
sysrepo gets built with `-pedantic`, meaning that libyang causes extra
warnings there.

Note that we still cannot enable `-pedantic` for libyang itself; some
extra warnings are still being reported:

- casts between `void*` and function pointers,
- forward references to enums,
- void pointer arithmetics,
- `return` in a function returning void,
- zero-sized arrays,
- string literals longer than 4kB,
- empty initializer lists,
- macros with zero variadic arguments.