XPath: deref(): improve error messages
While the error message was accurate, it was not as helpful as it could
be. Before:

  libyang warn: Argument #1 of xpath_deref is a list node "amplifier".

Now:

  libyang warn: Argument #1 of deref() is a list node "/tip-photonic-equipment:amplifier". Expected "leafref" or "instance-identifier".

Internally, the error handling can be simplified because we are now
printing the exact same thing in both cases. The user probably doesn't
care whether deref() hit a schema non-terminal or a terminal which is of
a wrong type anyway.

Unfortunately, it seems that libyang doesn't track the source location
of these XPaths. That's a pity especially in case of `yanglint` where
other tools, such as pyang, can provide a file name + line number. That
makes it much, much easier to map error messages to source code
locations. And since this function does not have access to the outer
enclosing element, we cannot print its XPath either.

Fixes #1710